requirements.txt
Django>=2.1.2,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0  # handles comms between django and postgres.
pylint>=2.4.4,<2.5.0
